requirements.txt
# FastAPI framework
fastapi>=0.110,<1.0

# ASGI server
uvicorn[standard]>=0.29,<1.0

# Templating
jinja2>=3.1

# Environment variables
python-dotenv>=1.0

# Password hashing
passlib[bcrypt]>=1.7

# SQLAlchemy ORM
sqlalchemy>=1.4,<2.1

# PostgreSQL adapter
psycopg2-binary>=2.9

# JWT handling
python-jose[cryptography]>=3.3

# Pydantic with email validator
pydantic[email]>=1.10,<3.0

# HTTP requests
requests>=2.31

# Data analysis
pandas>=2.0

# Google OAuth and API clients
google-auth>=2.29
google-auth-oauthlib>=1.2
google-api-python-client>=2.127

# Session signing / secure cookies
itsdangerous>=2.1

# For handling multipart/form-data
python-multipart>=0.0.6

